Make the visitor of the spatial triangle tree a public interface
