+++ x-pack/plugin/spatial/src/main/java/org/elasticsearch/xpack/spatial/index/fielddata/TriangleTreeReader.java
@@ -15,7 +15,7 @@ import java.io.IOException;
  * A tree reader for a previous serialized {@link org.elasticsearch.geometry.Geometry} using
  * {@link TriangleTreeWriter}.
  *
- * The tree structure is navigated using a {@link Visitor}.
+ * The tree structure is navigated using a {@link TriangleTreeVisitor}.
  *
  */
 class TriangleTreeReader {
@@ -23,16 +23,15 @@ class TriangleTreeReader {
     private TriangleTreeReader() {}
 
     /**
-     * Visit the Triangle tree using the {@link Visitor} provided.
+     * Visit the Triangle tree using the {@link TriangleTreeVisitor} provided.
      */
-    public static void visit(ByteArrayStreamInput input, TriangleTreeReader.Visitor visitor, int thisMaxX, int thisMaxY)
-        throws IOException {
+    public static void visit(ByteArrayStreamInput input, TriangleTreeVisitor visitor, int thisMaxX, int thisMaxY) throws IOException {
         visit(input, visitor, true, thisMaxX, thisMaxY, true);
     }
 
     private static boolean visit(
         ByteArrayStreamInput input,
-        TriangleTreeReader.Visitor visitor,
+        TriangleTreeVisitor visitor,
         boolean splitX,
         int thisMaxX,
         int thisMaxY,
@@ -90,13 +89,8 @@ class TriangleTreeReader {
         return visitor.push();
     }
 
-    private static boolean pushLeft(
-        ByteArrayStreamInput input,
-        TriangleTreeReader.Visitor visitor,
-        int thisMaxX,
-        int thisMaxY,
-        boolean splitX
-    ) throws IOException {
+    private static boolean pushLeft(ByteArrayStreamInput input, TriangleTreeVisitor visitor, int thisMaxX, int thisMaxY, boolean splitX)
+        throws IOException {
         int nextMaxX = Math.toIntExact(thisMaxX - input.readVLong());
         int nextMaxY = Math.toIntExact(thisMaxY - input.readVLong());
         int size = input.readVInt();
@@ -110,7 +104,7 @@ class TriangleTreeReader {
 
     private static boolean pushRight(
         ByteArrayStreamInput input,
-        TriangleTreeReader.Visitor visitor,
+        TriangleTreeVisitor visitor,
         int thisMaxX,
         int thisMaxY,
         int thisMinX,
@@ -131,126 +125,5 @@ class TriangleTreeReader {
             input.skipBytes(rightSize);
         }
         return visitor.push();
-    }
-
-    /** Visitor for triangle interval tree */
-    interface Visitor {
-
-        /** visit a node point. */
-        void visitPoint(int x, int y);
-
-        /** visit a node line. */
-        void visitLine(int aX, int aY, int bX, int bY, byte metadata);
-
-        /** visit a node triangle. */
-        void visitTriangle(int aX, int aY, int bX, int bY, int cX, int cY, byte metadata);
-
-        /** Should the visitor keep visiting the tree. Called after visiting a node or skipping
-         * a tree branch, if the return value is {@code false}, no more nodes will be visited. */
-        boolean push();
-
-        /** Should the visitor visit nodes that have bounds greater or equal
-         * than the {@code minX} provided. */
-        boolean pushX(int minX);
-
-        /** Should the visitor visit nodes that have bounds greater or equal
-         * than the {@code minY} provided. */
-        boolean pushY(int minY);
-
-        /** Should the visitor visit nodes that have bounds lower or equal than the
-         * {@code maxX} and {@code minX} provided. */
-        boolean push(int maxX, int maxY);
-
-        /** Should the visitor visit the tree given the bounding box of the tree. Called before
-         * visiting the tree. */
-        boolean push(int minX, int minY, int maxX, int maxY);
-    }
-
-    /** Visitor for triangle interval tree which decodes the coordinates */
-    public abstract static class DecodedVisitor implements Visitor {
-
-        private final CoordinateEncoder encoder;
-
-        DecodedVisitor(CoordinateEncoder encoder) {
-            this.encoder = encoder;
-        }
-
-        @Override
-        public void visitPoint(int x, int y) {
-            visitDecodedPoint(encoder.decodeX(x), encoder.decodeY(y));
-        }
-
-        /**
-         * Equivalent to {@link #visitPoint(int, int)} but coordinates are decoded.
-         */
-        abstract void visitDecodedPoint(double x, double y);
-
-        @Override
-        public void visitLine(int aX, int aY, int bX, int bY, byte metadata) {
-            visitDecodedLine(encoder.decodeX(aX), encoder.decodeY(aY), encoder.decodeX(bX), encoder.decodeY(bY), metadata);
-        }
-
-        /**
-         * Equivalent to {@link #visitLine(int, int, int, int, byte)} but coordinates are decoded.
-         */
-        abstract void visitDecodedLine(double aX, double aY, double bX, double bY, byte metadata);
-
-        @Override
-        public void visitTriangle(int aX, int aY, int bX, int bY, int cX, int cY, byte metadata) {
-            visitDecodedTriangle(
-                encoder.decodeX(aX),
-                encoder.decodeY(aY),
-                encoder.decodeX(bX),
-                encoder.decodeY(bY),
-                encoder.decodeX(cX),
-                encoder.decodeY(cY),
-                metadata
-            );
-        }
-
-        /**
-         * Equivalent to {@link #visitTriangle(int, int, int, int, int, int, byte)} but coordinates are decoded.
-         */
-        abstract void visitDecodedTriangle(double aX, double aY, double bX, double bY, double cX, double cY, byte metadata);
-
-        @Override
-        public boolean pushX(int minX) {
-            return pushDecodedX(encoder.decodeX(minX));
-        }
-
-        /**
-         * Equivalent to {@link #pushX(int)}  but coordinates are decoded.
-         */
-        abstract boolean pushDecodedX(double minX);
-
-        @Override
-        public boolean pushY(int minY) {
-            return pushDecodedY(encoder.decodeY(minY));
-        }
-
-        /**
-         * Equivalent to {@link #pushY(int)}  but coordinates are decoded.
-         */
-        abstract boolean pushDecodedY(double minX);
-
-        @Override
-        public boolean push(int maxX, int maxY) {
-            return pushDecoded(encoder.decodeX(maxX), encoder.decodeY(maxY));
-        }
-
-        /**
-         * Equivalent to {@link #push(int, int)} but coordinates are decoded.
-         */
-        abstract boolean pushDecoded(double maxX, double maxY);
-
-        @Override
-        public boolean push(int minX, int minY, int maxX, int maxY) {
-            return pushDecoded(encoder.decodeX(minX), encoder.decodeY(minY), encoder.decodeX(maxX), encoder.decodeY(maxY));
-        }
-
-        /**
-         * Equivalent to {@link #push(int, int, int, int)} but coordinates are decoded.
-         */
-        abstract boolean pushDecoded(double minX, double minY, double maxX, double maxY);
     }
 }
