+++ x-pack/plugin/spatial/src/test/java/org/elasticsearch/xpack/spatial/index/fielddata/TriangleTreeTests.java
@@ -42,7 +42,7 @@ public class TriangleTreeTests extends ESTestCase {
         assertThat(fieldList.size(), equalTo(visitor.counter));
     }
 
-    private static class TriangleCounterVisitor implements TriangleTreeReader.Visitor {
+    private static class TriangleCounterVisitor implements TriangleTreeVisitor {
 
         int counter;
